secure-boot.nix: Disable systemd initrd
Disable systemd as initrd to get impermanence back. Lose TPM unlock
in the process. systemd as initrd explicitly does not support
`pre*Commands` or `post*Commands` as stated in the PR introducing it:

https://github.com/NixOS/nixpkgs/pull/164943

This is evident when trying to make apply changes using those
declarations, and resulting in the same generation/outputs. That's
because nothing is actually changing. The scripts are never run.